// Client/src/api/Api.ts
export interface CategoryResponse {
  /** @format int32 */
  id?: number;
  name?: string;
}

export interface CreateCategoryRequest {
  name?: string;
}

export interface ListingResponse {
  vendorName?: string;
  categoryName?: string;
  /** @format int32 */
  listingId?: number;
  title?: string;
  description?: string;
  /** @format decimal */
  price?: number;
  /** @format int32 */
  stock?: number;
  /** @format int32 */
  categoryId?: number;
  /** @format int32 */
  vendorId?: number;
}

export interface CreateListingRequest {
  title?: string;
  description?: string;
  /** @format decimal */
  price?: number;
  /** @format int32 */
  stock?: number;
  /** @format int32 */
  categoryId?: number;
}

export interface UpdateListingRequest {
  /** @format int32 */
  listingIdForLookup?: number;
  title?: string;
  description?: string;
  /** @format decimal */
  price?: number;
  /** @format int32 */
  stock?: number;
  /** @format int32 */
  categoryId?: number;
}

export interface OrderResponse {
  /** @format int32 */
  id?: number;
  /** @format int32 */
  listingId?: number;
  productTitle?: string;
  /** @format int32 */
  quantity?: number;
  /** @format decimal */
  pricePaid?: number;
  /** @format date-time */
  createdAt?: string;
}

export interface CreateOrderRequest {
  /** @format int32 */
  listingId?: number;
  /** @format int32 */
  quantity?: number;
}

export interface UserResponse {
  /** @format int32 */
  userId?: number;
  userName?: string;
}

export type QueryParamsType = Record<string | number, any>;
export type ResponseFormat = keyof Omit<Body, "body" | "bodyUsed">;

export interface FullRequestParams extends Omit<RequestInit, "body"> {
  /** set parameter to `true` for call `securityWorker` for this request */
  secure?: boolean;
  /** request path */
  path: string;
  /** content type of request body */
  type?: ContentType;
  /** query params */
  query?: QueryParamsType;
  /** format of response (i.e. response.json() -> format: "json") */
  format?: ResponseFormat;
  /** request body */
  body?: unknown;
  /** base url */
  baseUrl?: string;
  /** request cancellation token */
  cancelToken?: CancelToken;
}

export type RequestParams = Omit<
  FullRequestParams,
  "body" | "method" | "query" | "path"
>;
// ...
export enum ContentType {
  Json = "application/json",
  JsonApi = "application/vnd.api+json",
  FormData = "multipart/form-data",
  UrlEncoded = "application/x-www-form-urlencoded",
  Text = "text/plain",
}
// ...
export class Api<
  SecurityDataType extends unknown,
> extends HttpClient<SecurityDataType> {
  getCategories = {
    /**
     * No description
     *
     * @tags Category
     * @name CategoryGetCategories
     * @request GET:/GetCategories
     */
    categoryGetCategories: (params: RequestParams = {}) =>
      this.request<CategoryResponse[], any>({
        path: `/GetCategories`,
        method: "GET",
        format: "json",
        ...params,
      }),
  };
  createCategory = {
    /**
     * No description
     *
     * @tags Category
     * @name CategoryCreateCategory
     * @request POST:/CreateCategory
     */
    categoryCreateCategory: (
      data: CreateCategoryRequest,
      params: RequestParams = {},
    ) =>
      this.request<CategoryResponse, any>({
        path: `/CreateCategory`,
        method: "POST",
        body: data,
        type: ContentType.Json,
        format: "json",
        ...params,
      }),
  };
  deleteCategory = {
    /**
     * No description
     *
     * @tags Category
     * @name CategoryDeleteCategory
     * @request DELETE:/DeleteCategory
     */
    categoryDeleteCategory: (
      query?: {
        /** @format int32 */
        categoryId?: number;
      },
      params: RequestParams = {},
    ) =>
      this.request<void, any>({
        path: `/DeleteCategory`,
        method: "DELETE",
        query: query,
        ...params,
      }),
  };
  getListings = {
    /**
     * No description
     *
     * @tags Listing
     * @name ListingGetListings
     * @request GET:/GetListings
     */
    listingGetListings: (
      query?: {
        /** @format int32 */
        categoryId?: number | null;
      },
      params: RequestParams = {},
    ) =>
      this.request<ListingResponse[], any>({
        path: `/GetListings`,
        method: "GET",
        query: query,
        format: "json",
        ...params,
      }),
  };
  createListing = {
    /**
     * No description
     *
     * @tags Listing
     * @name ListingCreateListing
     * @request POST:/CreateListing
     */
    listingCreateListing: (
      data: CreateListingRequest,
      params: RequestParams = {},
    ) =>
      this.request<ListingResponse, any>({
        path: `/CreateListing`,
        method: "POST",
        body: data,
        type: ContentType.Json,
        format: "json",
        ...params,
      }),
  };
  updateListing = {
    /**
     * No description
     *
     * @tags Listing
     * @name ListingUpdateListing
     * @request PUT:/UpdateListing
     */
    listingUpdateListing: (
      data: UpdateListingRequest,
      params: RequestParams = {},
    ) =>
      this.request<ListingResponse, any>({
        path: `/UpdateListing`,
        method: "PUT",
        body: data,
        type: ContentType.Json,
        format: "json",
        ...params,
      }),
  };
  deleteListings = {
    /**
     * No description
     *
     * @tags Listing
     * @name ListingDeleteListings
     * @request DELETE:/DeleteListings
     */
    listingDeleteListings: (
      query?: {
        /** @format int32 */
        listingId?: number;
      },
      params: RequestParams = {},
    ) =>
      this.request<void, any>({
        path: `/DeleteListings`,
        method: "DELETE",
        query: query,
        ...params,
      }),
  };
  placeOrder = {
    /**
     * No description
     *
     * @tags Order
     * @name OrderPlaceOrder
     * @request POST:/PlaceOrder
     */
    orderPlaceOrder: (data: CreateOrderRequest, params: RequestParams = {}) =>
      this.request<OrderResponse, any>({
        path: `/PlaceOrder`,
        method: "POST",
        body: data,
        type: ContentType.Json,
        format: "json",
        ...params,
      }),
  };
  getMyListings = {
    /**
     * No description
     *
     * @tags User
     * @name UserGetMyListings
     * @request GET:/GetMyListings
     */
    userGetMyListings: (params: RequestParams = {}) =>
      this.request<ListingResponse[], any>({
        path: `/GetMyListings`,
        method: "GET",
        format: "json",
        ...params,
      }),
  };
  getMe = {
    /**
     * No description
     *
     * @tags User
     * @name UserGetMe
     * @request GET:/GetMe
     */
    userGetMe: (params: RequestParams = {}) =>
      this.request<UserResponse, any>({
        path: `/GetMe`,
        method: "GET",
        format: "json",
        ...params,
      }),
  };
}